feat: 접근자 생성 메소드 분리
